tidied up - removed some unused files & config settings, moved spark session builder to function
