+++ config.toml
@@ -1,22 +1,7 @@
-project_name = "Example-Publication"
+project_name = "example_pipeline_pyspark_version"
 
-server = ''                   
-database = ''          
-schema = ''
-table = ''  
 data_url = "https://s3.eu-west-2.amazonaws.com/files.digital.nhs.uk/assets/Services/Artificial+data/Artificial+HES+final/artificial_hes_ae_202302_v1_sample.zip"
-data_hes_path = "data_in/artificial_hes_ae_202302_v1_sample.zip/artificial_hes_ae_202302_v1_sample/artificial_hes_ae_2122.csv"
-
-# The above variables reference an example location in SQL Server, they will need to be changed for your project
-# If you are using Databricks instead of SQL Server you will not need to set the 'server' variable
-# Your project may use more than 1 table, you should add more table names and assignments as needed
-
-# Include any other configuration parameters, e.g. dates
-start_date = "1970-01-01"
-end_date = "2022-01-01"
-
-# Fill missing value for specific column
-filled_value = 2022
+path_to_downloaded_data = "data_in/artificial_hes_ae_202302_v1_sample.zip/artificial_hes_ae_202302_v1_sample/artificial_hes_ae_2122.csv"
 
 # Here we describe where the output and logs are saved, change as necessary
 output_dir = ''
