--- sql/example.sql
@@ -1,2 +1,0 @@
-SELECT NAME, ADDRESS_LINE_5, OPEN_DATE, CLOSE_DATE 
-FROM {database}.{schema}.{table};
