+++ src/utils/spark.py
@@ -1,0 +1,10 @@
+from pyspark.sql import SparkSession
+
+def create_spark_session(app_name):
+    spark_session = (SparkSession
+        .builder
+        .appName(app_name)
+        .getOrCreate()
+    )
+
+    return spark_session
